Related work items: #12722, #15249, #15250

+++ dotnet/corecomponents/net.atos.daf.ct2.dashboard/net.atos.daf.ct2.dashboard/entity/TodayLiveVehicle.cs
@@ -12,12 +12,12 @@ namespace net.atos.daf.ct2.dashboard.entity
         public int DriverCount { get; set; }
         public int TodayActiveVinCount { get; set; }
         public long TodayTimeBasedUtilizationRate { get; set; }
-        public int TodayDistanceBasedUtilization { get; set; }
+        public long TodayDistanceBasedUtilization { get; set; }
         public int CriticleAlertCount { get; set; }
         public string YesterdayVin { get; set; }
         public int YesterdayActiveVinCount { get; set; }
         public long YesterDayTimeBasedUtilizationRate { get; set; }
-        public int YesterDayDistanceBasedUtilization { get; set; }
+        public long YesterDayDistanceBasedUtilization { get; set; }
     }
 
     public class TodayLiveVehicleRequest
